@@ -13,27 +13,27 @@
     "start:ssr": "npm run build:ssr && npm run serve:ssr",
     "dev:ssr": "ng run stroyka:serve-ssr",
     "serve:ssr": "node dist/stroyka/server/main.js",
-    "build:ssr": "ng build --prod && ng run stroyka:server:production",
+    "build:ssr": "ng build --configuration production && ng run stroyka:server:production",
     "prerender": "ng run stroyka:prerender"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~11.2.7",
-    "@angular/cdk": "^10.2.5",
-    "@angular/common": "~10.0.0",
-    "@angular/compiler": "~11.2.7",
-    "@angular/core": "~11.2.7",
-    "@angular/forms": "~11.2.7",
-    "@angular/material": "^10.2.5",
-    "@angular/platform-browser": "~11.2.7",
-    "@angular/platform-browser-dynamic": "~11.2.7",
-    "@angular/platform-server": "~11.2.7",
-    "@angular/router": "~11.2.7",
-    "@angular/service-worker": "~11.2.7",
+    "@angular/animations": "^14.3.0",
+    "@angular/cdk": "^14.2.7",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/material": "^14.2.7",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/platform-server": "^14.3.0",
+    "@angular/router": "^14.3.0",
+    "@angular/service-worker": "^14.3.0",
     "@aspnet/signalr": "^1.0.26",
     "@fortawesome/fontawesome-free": "^5.13.1",
     "@m0t0r/ngx-slider": "^1.0.0",
-    "@nguniversal/express-engine": "^10.0.0",
+    "@nguniversal/express-engine": "^7.0.1",
     "bootstrap": "^4.5.0",
     "crypto-js": "^4.2.0",
     "express": "^4.19.2",
@@ -46,14 +46,14 @@
     "rxjs": "~6.5.5",
     "secure-web-storage": "^1.0.2",
     "tslib": "^2.0.0",
-    "zone.js": "~0.10.3"
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1000.0",
-    "@angular/cli": "^11.2.6",
-    "@angular/compiler-cli": "~11.2.7",
-    "@angular/language-service": "~11.2.7",
-    "@nguniversal/builders": "^10.0.0",
+    "@angular-devkit/build-angular": "^14.2.13",
+    "@angular/cli": "^14.2.13",
+    "@angular/compiler-cli": "^14.3.0",
+    "@angular/language-service": "^14.3.0",
+    "@nguniversal/builders": "^14.2.3",
     "@types/crypto-js": "^3.1.47",
     "@types/express": "^4.17.0",
     "@types/jasmine": "~3.5.11",
@@ -65,7 +65,7 @@
     "eslint": "^7.16.0",
     "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.0.0",
+    "karma": "~6.4.4",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~3.3.0",
@@ -74,7 +74,7 @@
     "svg-sprite": "^1.5.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.2",
-    "typescript": "~4.1.5"
+    "typescript": "~4.6.4"
   },
   "browser": {
     "crypto": false
